+++ velox/experimental/stateful/state/RocksDBKeyedStateBackend.h
@@ -58,12 +58,12 @@ class RocksDBKeyedStateBackend : public KeyedStateBackend {
   std::shared_ptr<MapState<uint32_t, int, uint32_t, RowVectorPtr>>
   getOrCreateRankMapState(StateDescriptor& stateDescriptor) override;
 
-  std::shared_ptr<InternalTimerService<uint32_t, int64_t>> createTimerService(
-      Triggerable<uint32_t, int64_t>* triggerable) override;
+  std::shared_ptr<InternalTimerService<int64_t, int64_t>> createTimerService(
+      Triggerable<int64_t, int64_t>* triggerable) override;
 
-  std::shared_ptr<InternalTimerService<uint32_t, TimeWindow>>
+  std::shared_ptr<InternalTimerService<int64_t, TimeWindow>>
   createGroupWindowAggTimerService(
-      Triggerable<uint32_t, TimeWindow>* triggerable) override;
+      Triggerable<int64_t, TimeWindow>* triggerable) override;
 
   // Deprecated, this maybe removed later
   void snapshot(
